Add Joukowsky and Karman-Trefftz transforms

+++ code/AirfoilTools/point_maker.hpp
@@ -8,10 +8,11 @@
 namespace AirfoilTools::PointMaker {
 
 /// @brief generates a circle
-/// @param dx
+/// this circle will intersect the point 1 + 0i
+/// @param dx the scale in the x direction, must be less than 0
 /// @param dy
 /// @param pointcount
 /// @return a vector of complex numbers
-complex_vector make_points(double dx, double dy, double pointcount);
+complex_vector make_points(double dx, double dy, int pointcount);
 
 } // namespace AirfoilTools::PointMaker
